Monitoring_Agent/requirements.txt
playwright==1.49.1
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
numpy>=1.26.0
pydantic>=2.0.0
aiofiles>=24.1.0
groq>=0.13.0
